deploy-yolo/requirements.txt
fastapi==0.119.1
uvicorn==0.30.6
pillow==10.4.0
ultralytics==8.3.202

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
torchaudio==2.5.1+cpu
